[NOID] cleanup and fix tests
